+++ src/Utils/SalesChannelUtils.php
@@ -9,18 +9,10 @@ use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
-use Shopware\Core\System\Country\CountryCollection;
-use Shopware\Core\System\Country\CountryEntity;
 use Shopware\Core\System\Currency\CurrencyCollection;
 use Shopware\Core\System\Currency\CurrencyEntity;
-use Shopware\Core\System\Language\LanguageCollection;
-use Shopware\Core\System\Language\LanguageEntity;
-use Shopware\Core\System\Locale\LocaleCollection;
-use Shopware\Core\System\Locale\LocaleEntity;
 use Shopware\Core\System\SalesChannel\SalesChannelCollection;
 use Shopware\Core\System\SalesChannel\SalesChannelEntity;
-use Shopware\Core\System\Snippet\Aggregate\SnippetSet\SnippetSetCollection;
-use Shopware\Core\System\Snippet\Aggregate\SnippetSet\SnippetSetEntity;
 use Shopware\Core\System\Tax\TaxCollection;
 use Shopware\Core\System\Tax\TaxEntity;
 
@@ -37,21 +29,13 @@ readonly class SalesChannelUtils
 {
     /**
      * @param EntityRepository<SalesChannelCollection> $salesChannelRepository
-     * @param EntityRepository<SnippetSetCollection>   $snippetSetRepository
      * @param EntityRepository<TaxCollection>          $taxRepository
-     * @param EntityRepository<CountryCollection>      $countryRepository
-     * @param EntityRepository<LanguageCollection>     $languageRepository
      * @param EntityRepository<CurrencyCollection>     $currencyRepository
-     * @param EntityRepository<LocaleCollection>       $localeRepository
      */
     public function __construct(
         private EntityRepository $salesChannelRepository,
-        private EntityRepository $snippetSetRepository,
         private EntityRepository $taxRepository,
-        private EntityRepository $countryRepository,
-        private EntityRepository $languageRepository,
         private EntityRepository $currencyRepository,
-        private EntityRepository $localeRepository,
     ) {
     }
 
@@ -108,64 +92,6 @@ readonly class SalesChannelUtils
             ->first();
 
         return $currency instanceof CurrencyEntity ? $currency : null;
-    }
-
-    // TODO: Move to LanguageAndLocaleUtils
-    public function getLanguage(string $languageName): ?LanguageEntity
-    {
-        $criteria = (new Criteria())->addFilter(
-            new EqualsFilter('name', $languageName),
-        )->setLimit(1);
-
-        $language = $this->languageRepository
-            ->search($criteria, Context::createDefaultContext())
-            ->first();
-
-        return $language instanceof LanguageEntity ? $language : null;
-    }
-
-    // TODO: Move to LanguageAndLocaleUtils
-    public function getLocale(string $code): ?LocaleEntity
-    {
-        $criteria = (new Criteria())->addFilter(
-            new EqualsFilter('code', $code),
-        )->setLimit(1);
-
-        $locale = $this->localeRepository
-            ->search($criteria, Context::createDefaultContext())
-            ->first();
-
-        return $locale instanceof LocaleEntity ? $locale : null;
-    }
-
-    // TODO: Move to LanguageAndLocaleUtils
-    public function getCountry(string $countryIso): ?CountryEntity
-    {
-        $criteria = (new Criteria())->addFilter(
-            new EqualsFilter('iso', $countryIso),
-        )->setLimit(1);
-
-        $country = $this->countryRepository
-            ->search(
-                $criteria,
-                Context::createDefaultContext(),
-            )->first();
-
-        return $country instanceof CountryEntity ? $country : null;
-    }
-
-    // TODO: Move to LanguageAndLocaleUtils
-    public function getSnippetSet(string $countryCodeIso): ?SnippetSetEntity
-    {
-        $criteria = (new Criteria())->addFilter(
-            new EqualsFilter('iso', $countryCodeIso),
-        )->setLimit(1);
-
-        $snippetSet = $this->snippetSetRepository
-            ->search($criteria, Context::createDefaultContext())
-            ->first();
-
-        return $snippetSet instanceof SnippetSetEntity ? $snippetSet : null;
     }
 
     // TODO: Move to TaxUtils