+++ CHANGELOG.md
@@ -24,15 +24,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `FixtureTrait::runFixtureGroup` is a new function to execute whole fixture groups with optionally dependencies
 - Each fixture now has direct access to the FixtureHelper using `$this->helper`
   - **Breaking** If you have the helper (or any other helper) previously assigned to `$this->helper` it will either fail or override the FixturePlugin helper
+- **Breaking** Moved `SalesChannelUtils::getLanguage()` to `LanguageAndLocaleUtils::getLanguage()`
+- **Breaking** Moved `SalesChannelUtils::getLocale()` to `LanguageAndLocaleUtils::getLocale()`
+- **Breaking** Moved `SalesChannelUtils::getCountry()` to `LanguageAndLocaleUtils::getCountry()`
+- **Breaking** Moved `SalesChannelUtils::getSnippetSet()` to `LanguageAndLocaleUtils::getSnippetSet()`
 
 ### Removed
 - Dropped support for PHP 8.1
 - Dropped support for Shopware 6.3 & 6.4
-- Removed FixtureBag
-- CategoryUtils
+- **Breaking** Removed FixtureBag
+- **Breaking** CategoryUtils
   - Removed method `getFirst` on CategoryUtils
   - Removed method `getByName` on CategoryUtils
-- Renamed `CategoryUtils` to `SalutationUtils`
+- **Breaking** Renamed `CategoryUtils` to `SalutationUtils`
 
 ## [2.4.0] - 2023-11-15
 ### Added